refactor: arrival allocations to own service
