+++ app/Providers/StandServiceProvider.php
@@ -7,6 +7,7 @@ use App\Allocator\Stand\CargoFlightPreferredArrivalStandAllocator;
 use App\Allocator\Stand\CargoFlightArrivalStandAllocator;
 use App\Allocator\Stand\CidReservedArrivalStandAllocator;
 use App\Services\Stand\AirfieldStandService;
+use App\Services\Stand\ArrivalAllocationService;
 use App\Services\Stand\StandAdminService;
 use App\Services\Stand\StandAssignmentsService;
 use App\Services\Stand\StandService;
@@ -30,6 +31,14 @@ class StandServiceProvider extends ServiceProvider
     {
         $this->app->singleton(StandService::class, function (Application $application) {
             return new StandService(
+                $application->make(StandAssignmentsService::class),
+                $application->make(AirfieldStandService::class)
+            );
+        });
+
+        $this->app->singleton(ArrivalAllocationService::class, function (Application $application) {
+            return new ArrivalAllocationService(
+                $application->make(StandAssignmentsService::class),
                 [
                     $application->make(CidReservedArrivalStandAllocator::class),
                     $application->make(CallsignFlightplanReservedArrivalStandAllocator::class),
@@ -42,12 +51,9 @@ class StandServiceProvider extends ServiceProvider
                     $application->make(CargoAirlineFallbackStandAllocator::class),
                     $application->make(DomesticInternationalStandAllocator::class),
                     $application->make(FallbackArrivalStandAllocator::class),
-                ],
-                $application->make(StandAssignmentsService::class),
-                $application->make(AirfieldStandService::class)
+                ]
             );
         });
-
         $this->app->singleton(StandReservationsImport::class);
         $this->app->singleton(StandAdminService::class);
         $this->app->singleton(AirfieldStandService::class);
