+++ tests/app/Services/Stand/StandServiceTest.php
@@ -378,226 +378,6 @@ class StandServiceTest extends BaseFunctionalTestCase
         $this->assertNull($this->dependency->updated_at);
     }
 
-
-
-    public function testItHasAllocatorPreference()
-    {
-        $this->assertEquals(
-            [
-                CidReservedArrivalStandAllocator::class,
-                CallsignFlightplanReservedArrivalStandAllocator::class,
-                CargoFlightPreferredArrivalStandAllocator::class,
-                CargoFlightArrivalStandAllocator::class,
-                AirlineCallsignSlugArrivalStandAllocator::class,
-                AirlineDestinationArrivalStandAllocator::class,
-                AirlineArrivalStandAllocator::class,
-                AirlineTerminalArrivalStandAllocator::class,
-                CargoAirlineFallbackStandAllocator::class,
-                DomesticInternationalStandAllocator::class,
-                FallbackArrivalStandAllocator::class,
-            ],
-            $this->service->getAllocatorPreference()
-        );
-    }
-
-    public function testItAllocatesAStandFromAllocator()
-    {
-        $this->expectsEvents(StandAssignedEvent::class);
-        StandReservation::create(
-            [
-                'callsign' => 'BMI221',
-                'stand_id' => 1,
-                'start' => Carbon::now()->subMinute(),
-                'end' => Carbon::now()->addMinute(),
-                'destination' => 'EGLL',
-                'origin' => 'EGSS',
-            ]
-        );
-
-        $aircraft = NetworkAircraftService::createOrUpdateNetworkAircraft(
-            'BMI221',
-            [
-                'planned_aircraft' => 'B738',
-                'planned_destairport' => 'EGLL',
-                'planned_depairport' => 'EGSS',
-                'groundspeed' => 150,
-                // London
-                'latitude' => 51.487202,
-                'longitude' => -0.466667,
-            ]
-        );
-
-        $assignment = $this->service->allocateStandForAircraft($aircraft);
-        $this->assertEquals(1, $assignment->stand_id);
-        $this->assertEquals('BMI221', $assignment->callsign);
-    }
-
-    public function testItDoesntAllocateStandIfPerformingCircuits()
-    {
-        $this->doesntExpectEvents(StandAssignedEvent::class);
-        $aircraft = NetworkAircraftService::createOrUpdateNetworkAircraft(
-            'BMI221',
-            [
-                'planned_aircraft' => 'B738',
-                'planned_destairport' => 'EGLL',
-                'planned_depairport' => 'EGLL',
-                'groundspeed' => 150,
-                // London
-                'latitude' => 51.487202,
-                'longitude' => -0.466667,
-            ]
-        );
-
-        $this->assertNull($this->service->allocateStandForAircraft($aircraft));
-        $this->assertFalse(StandAssignment::where('callsign', 'BMI221')->exists());
-    }
-
-    public function testItDoesntPerformAllocationIfStandTooFarFromAirfield()
-    {
-        $this->doesntExpectEvents(StandAssignedEvent::class);
-        $aircraft = NetworkAircraftService::createOrUpdateNetworkAircraft(
-            'BMI221',
-            [
-                'planned_aircraft' => 'B738',
-                'planned_destairport' => 'EGLL',
-                'groundspeed' => 100,
-                // Lambourne
-                'latitude' => 51.646099,
-                'longitude' => 0.151667,
-            ]
-        );
-
-        $this->assertNull($this->service->allocateStandForAircraft($aircraft));
-        $this->assertFalse(StandAssignment::where('callsign', 'BMI221')->exists());
-    }
-
-    public function testItDoesntPerformAllocationIfAircraftHasNoGroundspeed()
-    {
-        $this->doesntExpectEvents(StandAssignedEvent::class);
-        $aircraft = NetworkAircraftService::createOrUpdateNetworkAircraft(
-            'BMI221',
-            [
-                'planned_aircraft' => 'B738',
-                'planned_destairport' => 'EGLL',
-                'groundspeed' => 0,
-                // Lambourne
-                'latitude' => 51.646099,
-                'longitude' => 0.151667,
-            ]
-        );
-
-        $this->assertNull($this->service->allocateStandForAircraft($aircraft));
-        $this->assertFalse(StandAssignment::where('callsign', 'BMI221')->exists());
-    }
-
-    public function testItDoesntPerformAllocationIfNoStandAllocated()
-    {
-        // Delete all the stands so there's nothing to allocate
-        Stand::all()->each(function (Stand $stand) {
-            $stand->delete();
-        });
-
-        $this->doesntExpectEvents(StandAssignedEvent::class);
-        $aircraft = NetworkAircraftService::createOrUpdateNetworkAircraft(
-            'BMI221',
-            [
-                'planned_aircraft' => 'B738',
-                'planned_destairport' => 'EGLL',
-                'groundspeed' => 150,
-                // Lambourne
-                'latitude' => 51.646099,
-                'longitude' => 0.151667,
-            ]
-        );
-
-        $this->assertNull($this->service->allocateStandForAircraft($aircraft));
-        $this->assertFalse(StandAssignment::where('callsign', 'BMI221')->exists());
-    }
-
-    public function testItDoesntPerformAllocationIfStandAlreadyAssigned()
-    {
-        $this->doesntExpectEvents(StandAssignedEvent::class);
-        $aircraft = NetworkAircraftService::createOrUpdateNetworkAircraft(
-            'BMI221',
-            [
-                'planned_aircraft' => 'B738',
-                'planned_destairport' => 'EGLL',
-                'groundspeed' => 150,
-                // Lambourne
-                'latitude' => 51.646099,
-                'longitude' => 0.151667,
-            ]
-        );
-        StandAssignment::create(
-            [
-                'callsign' => 'BMI221',
-                'stand_id' => 1
-            ]
-        );
-
-        $this->assertNull($this->service->allocateStandForAircraft($aircraft));
-        $this->assertTrue(StandAssignment::where('callsign', 'BMI221')->where('stand_id', 1)->exists());
-    }
-
-    public function testItDoesntReturnAllocationIfAirfieldNotFound()
-    {
-        $this->doesntExpectEvents(StandAssignedEvent::class);
-        $aircraft = NetworkAircraftService::createOrUpdateNetworkAircraft(
-            'BMI221',
-            [
-                'planned_aircraft' => 'B738',
-                'planned_destairport' => 'EGXX',
-                'groundspeed' => 150,
-                // Lambourne
-                'latitude' => 51.646099,
-                'longitude' => 0.151667,
-            ]
-        );
-
-        $this->assertNull($this->service->allocateStandForAircraft($aircraft));
-        $this->assertFalse(StandAssignment::where('callsign', 'BMI221')->exists());
-    }
-
-    public function testItDoesntPerformAllocationIfUnknownAircraftType()
-    {
-        $this->doesntExpectEvents(StandAssignedEvent::class);
-        $aircraft = NetworkAircraftService::createOrUpdateNetworkAircraft(
-            'BMI221',
-            [
-                'planned_aircraft' => 'B736',
-                'planned_destairport' => 'EGLL',
-                'groundspeed' => 150,
-                // Lambourne
-                'latitude' => 51.646099,
-                'longitude' => 0.151667,
-            ]
-        );
-
-        $this->assertNull($this->service->allocateStandForAircraft($aircraft));
-        $this->assertFalse(StandAssignment::where('callsign', 'BMI221')->exists());
-    }
-
-    public function testItDoesntPerformAllocationIfAircraftTypeNotStandAssignable()
-    {
-        Aircraft::where('code', 'B738')->update(['allocate_stands' => false]);
-
-        $this->doesntExpectEvents(StandAssignedEvent::class);
-        $aircraft = NetworkAircraftService::createOrUpdateNetworkAircraft(
-            'BMI221',
-            [
-                'planned_aircraft' => 'B738',
-                'planned_destairport' => 'EGLL',
-                'groundspeed' => 150,
-                // Lambourne
-                'latitude' => 51.646099,
-                'longitude' => 0.151667,
-            ]
-        );
-
-        $this->assertNull($this->service->allocateStandForAircraft($aircraft));
-        $this->assertFalse(StandAssignment::where('callsign', 'BMI221')->exists());
-    }
-
     public function testItReturnsAnAircraftsStandAssignment()
     {
         $this->addStandAssignment('BAW959', 1);
